pyproject.toml
[project]
name = "pocarchitect"
version = "0.2.0"
description = "POC Architect - AI Agent that turns raw PoC URLs into weaponized Markdown blueprints for red teamers"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [{name = "RikterSkale"}]
keywords = ["red-team", "pentesting", "poc", "exploit-development", "ai-agent", "offsec", "cybersecurity"]

dependencies = [
    "typer[all]>=0.12.5",      # [all] brings shell completion support
    "rich>=13.7.0",
    "openai>=1.35.0",
    "python-dotenv>=1.0.0",
    "GitPython>=3.1.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
all = []   # kept for future extras / compatibility with Dockerfile/docs

[project.scripts]
pocarchitect = "pocarchitect.cli:app"

[project.urls]
Homepage = "https://github.com/rikterskale/POCArchitect-AI-Agent"
Repository = "https://github.com/rikterskale/POCArchitect-AI-Agent"
"Bug Tracker" = "https://github.com/rikterskale/POCArchitect-AI-Agent/issues"

[tool.setuptools.packages.find]
include = ["pocarchitect*"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
pocarchitect = ["*.md"]